[Feat] #32 - CompositionalLayoutFactory 추상화
